+++ compose.yml
@@ -50,7 +50,6 @@ services:
       - node_modules_ghost_magic-link:/home/ghost/ghost/magic-link/node_modules:delegated
       - node_modules_ghost_member-attribution:/home/ghost/ghost/member-attribution/node_modules:delegated
       - node_modules_ghost_member-events:/home/ghost/ghost/member-events/node_modules:delegated
-      - node_modules_ghost_members-api:/home/ghost/ghost/members-api/node_modules:delegated
       - node_modules_ghost_members-csv:/home/ghost/ghost/members-csv/node_modules:delegated
       - node_modules_ghost_milestones:/home/ghost/ghost/milestones/node_modules:delegated
       - node_modules_ghost_mw-error-handler:/home/ghost/ghost/mw-error-handler/node_modules:delegated
@@ -176,7 +175,6 @@ volumes:
   node_modules_ghost_magic-link: {}
   node_modules_ghost_member-attribution: {}
   node_modules_ghost_member-events: {}
-  node_modules_ghost_members-api: {}
   node_modules_ghost_members-csv: {}
   node_modules_ghost_milestones: {}
   node_modules_ghost_mw-error-handler: {}
