+++ ghost/core/core/server/services/members/api.js
@@ -1,7 +1,7 @@
 const stripeService = require('../stripe');
 const settingsCache = require('../../../shared/settings-cache');
 const settingsHelpers = require('../../services/settings-helpers');
-const MembersApi = require('@tryghost/members-api');
+const MembersApi = require('./members-api/members-api');
 const logging = require('@tryghost/logging');
 const mail = require('../mail');
 const models = require('../../models');
