+++ ghost/core/package.json
@@ -40,7 +40,7 @@
     "test:browser:setup": "npx playwright install",
     "test:ci:e2e": "c8 -c ./.c8rc.e2e.json -o coverage-e2e yarn test:e2e -b",
     "test:ci:regression": "yarn test:regression -b",
-    "test:ci:integration": "c8 -c ./.c8rc.e2e.json -o coverage-integration --lines 54 --functions 47 --branches 75 --statements 54 yarn test:integration -b",
+    "test:ci:integration": "c8 -c ./.c8rc.e2e.json -o coverage-integration --lines 53 --functions 47 --branches 74 --statements 53 yarn test:integration -b",
     "test:unit:slow": "yarn test:unit --reporter=mocha-slow-test-reporter",
     "test:int:slow": "yarn test:integration --reporter=mocha-slow-test-reporter",
     "test:e2e:slow": "yarn test:e2e --reporter=mocha-slow-test-reporter",
@@ -102,7 +102,6 @@
     "@tryghost/magic-link": "0.0.0",
     "@tryghost/member-attribution": "0.0.0",
     "@tryghost/member-events": "0.0.0",
-    "@tryghost/members-api": "0.0.0",
     "@tryghost/members-csv": "0.0.0",
     "@tryghost/members-offers": "0.0.0",
     "@tryghost/metrics": "1.0.37",
@@ -220,6 +219,7 @@
     "tiny-glob": "0.2.9",
     "tough-cookie": "4.1.4",
     "ua-parser-js": "1.0.40",
+    "uuid": "9.0.1",
     "xml": "1.0.1"
   },
   "optionalDependencies": {
@@ -233,7 +233,9 @@
     "@tryghost/webhook-mock-receiver": "0.2.14",
     "@types/bookshelf": "1.2.9",
     "@types/common-tags": "1.8.4",
+    "@types/jsonwebtoken": "9.0.6",
     "@types/node": "22.13.5",
+    "@types/node-jose": "1.1.13",
     "@types/sinon": "17.0.4",
     "@types/supertest": "6.0.3",
     "c8": "8.0.1",
@@ -245,6 +247,7 @@
     "find-root": "1.1.0",
     "form-data": "4.0.0",
     "inquirer": "8.2.6",
+    "jwk-to-pem": "2.0.7",
     "jwks-rsa": "3.2.0",
     "mocha": "10.8.2",
     "mocha-slow-test-reporter": "0.1.2",
