+++ ghost/core/test/unit/server/services/members/members-api/controllers/member/index.test.js
@@ -1,5 +1,5 @@
 const sinon = require('sinon');
-const MemberController = require('../../../../../lib/controllers/MemberController');
+const MemberController = require('../../../../../../../../core/server/services/members/members-api/controllers/MemberController');
 
 describe('MemberController', function () {
     describe('updateSubscription', function () {
