+++ ghost/core/test/unit/server/services/members/members-api/repositories/event.test.js
@@ -1,5 +1,5 @@
 const should = require('should');
-const EventRepository = require('../../../../lib/repositories/EventRepository');
+const EventRepository = require('../../../../../../../core/server/services/members/members-api/repositories/EventRepository');
 const sinon = require('sinon');
 const errors = require('@tryghost/errors');
