+++ ghost/core/test/unit/server/services/members/members-api/repositories/member.test.js
@@ -1,7 +1,7 @@
 const assert = require('assert/strict');
 const sinon = require('sinon');
 const DomainEvents = require('@tryghost/domain-events');
-const MemberRepository = require('../../../../lib/repositories/MemberRepository');
+const MemberRepository = require('../../../../../../../core/server/services/members/members-api/repositories/MemberRepository');
 const {SubscriptionCreatedEvent, OfferRedemptionEvent} = require('@tryghost/member-events');
 
 const mockOfferRedemption = {
