--- ghost/members-api/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@tryghost/members-api",
-  "version": "0.0.0",
-  "private": true,
-  "author": "Ghost Foundation",
-  "license": "MIT",
-  "main": "index.js",
-  "scripts": {
-    "dev": "echo \"Implement me!\"",
-    "test:unit": "NODE_ENV=testing c8 --all --reporter text --reporter cobertura -- mocha --reporter dot --reporter dot './test/**/*.test.js'",
-    "test": "yarn test:unit",
-    "lint:code": "eslint *.js lib/ --ext .js --cache",
-    "lint": "yarn lint:code && yarn lint:test",
-    "lint:test": "eslint -c test/.eslintrc.js test/ --ext .js --cache"
-  },
-  "files": [
-    "index.js",
-    "MembersAPI.js",
-    "lib",
-    "gateway"
-  ],
-  "devDependencies": {
-    "@tryghost/member-attribution": "0.0.0",
-    "@types/node-jose": "1.1.13",
-    "c8": "8.0.1",
-    "jwk-to-pem": "2.0.7",
-    "mocha": "10.8.2",
-    "nock": "13.5.6",
-    "should": "13.2.3",
-    "sinon": "15.2.0"
-  },
-  "dependencies": {
-    "@tryghost/domain-events": "0.0.0",
-    "@tryghost/errors": "1.3.5",
-    "@tryghost/logging": "2.4.21",
-    "@tryghost/magic-link": "0.0.0",
-    "@tryghost/member-events": "0.0.0",
-    "@tryghost/nql": "0.12.7",
-    "@tryghost/tpl": "0.1.32",
-    "@tryghost/validator": "0.2.14",
-    "@types/jsonwebtoken": "9.0.6",
-    "body-parser": "1.20.3",
-    "bson-objectid": "2.0.4",
-    "express": "4.21.2",
-    "got": "11.8.6",
-    "jsonwebtoken": "8.5.1",
-    "lodash": "4.17.21",
-    "moment": "2.29.4",
-    "node-jose": "2.2.0",
-    "uuid": "9.0.1"
-  }
-}
